move render elements function in separate module
